Split out restorePost
